Module Cobol_unit.Types

type 'a named_n_ordered = {

named : 'a Cobol_unit__.Unit_qualmap.TYPES.qualmap;

list : 'a list;

}

A type to manage sets of items in both named and ordered form. Note that in many cases anonymous elements of list many not belong to named.

type unit_config = {

unit_currency_signs : Cobol_common.Basics.CharSet.t;

unit_decimal_point : char;

}

Corresponds to the contents of the CONFIGURATION SECTION.

type data_definitions = {

data_items : Cobol_data.Types.data_definition named_n_ordered;

data_records : Cobol_data.Types.record list;

}

type procedure_paragraph = {

paragraph_name : Cobol_ptree.procedure_name Cobol_common.Srcloc.TYPES.with_loc option;

paragraph : Cobol_ptree.paragraph Cobol_common.Srcloc.TYPES.with_loc;

}

type procedure_section = {

section_name : Cobol_ptree.procedure_name Cobol_common.Srcloc.TYPES.with_loc;

section_paragraphs : procedure_paragraph Cobol_common.Srcloc.TYPES.with_loc named_n_ordered;

}

type procedure_block =

| Paragraph of procedure_paragraph Cobol_common.Srcloc.TYPES.with_loc

| Section of procedure_section Cobol_common.Srcloc.TYPES.with_loc

type procedure = procedure_block named_n_ordered

type cobol_unit = {

unit_name : string Cobol_common.Srcloc.TYPES.with_loc;

unit_parent_name : string Cobol_common.Srcloc.TYPES.with_loc option;

unit_config : unit_config;

unit_data : data_definitions;

unit_procedure : procedure;

}

Type of a main COBOL compilation unit

type t = cobol_unit Cobol_common.Srcloc.TYPES.with_loc

val block_name : procedure_block -> Cobol_ptree.qualname Cobol_common.Srcloc.TYPES.with_loc option

type group = Stdlib.Set.Make(Cobol_unit__.Unit_collections.M).t

type +'a group_map = 'a Cobol_unit__.Unit_collections.MAP.t